chore(RequestURLBuilder): move class to the main-package in 'implementation' subproject
