--- api/src/main/java/io/github/aivruu/repoviewer/repository/application/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Provides access to the API requests url-builder.
- *
- * @since 4.0.0
- */
-package io.github.aivruu.repoviewer.repository.application;
